+++ kamehouse-commons-core/src/main/java/com/nicobrest/kamehouse/commons/model/kamehousecommand/KameHouseShellScript.java
@@ -220,7 +220,7 @@ public abstract class KameHouseShellScript implements KameHouseCommand {
    * Override to set the command as a daemon.
    */
   @JsonIgnore
-  protected boolean isDaemon() {
+  public boolean isDaemon() {
     return false;
   }